+++ reviveme/api/v1/comment_controller.py
@@ -134,51 +134,3 @@ def comment_delete(comment_id):
     comment.deleted = True
     db.session.commit()
     return Response(status=200)
-
-@bp.route("/comments/<int:comment_id>/upvote", methods=["POST"])
-def comment_upvote(comment_id):
-    comment = db.get_or_404(Comment, comment_id)
-    if comment.deleted:
-        return Response(status=400, response=f"Comment with id {comment_id} has been deleted")
-    
-    data = UpVoteSchema().load(request.get_json())
-
-    vote = db.session.execute(
-        select(CommentVote).where(CommentVote.comment_id == comment_id, CommentVote.user_id == data['user_id'])
-    ).scalars().first()
-
-    if vote is None:
-        vote = CommentVote(user_id=data['user_id'], comment_id=comment_id, upvote=data['upvote'])
-        db.session.add(vote)
-    elif vote.upvote and not data['upvote']: # if user upvoted and now wants to remove upvote
-        db.session.delete(vote)
-    elif not vote.upvote and data['upvote']: # if user downvoted and now wants to upvote
-        vote.upvote = True
-        db.session.add(vote)
-    
-    db.session.commit()
-    return Response(status=200)
-
-@bp.route("/comments/<int:comment_id>/downvote", methods=["POST"])
-def comment_downvote(comment_id):
-    comment = db.get_or_404(Comment, comment_id)
-    if comment.deleted:
-        return Response(status=400, response=f"Comment with id {comment_id} has been deleted")
-    
-    data = DownVoteSchema().load(request.get_json())
-
-    vote = db.session.execute(
-        select(CommentVote).where(CommentVote.comment_id == comment_id, CommentVote.user_id == data['user_id'])
-    ).scalars().first()
-
-    if vote is None:
-        vote = CommentVote(user_id=data['user_id'], comment_id=comment_id, upvote=not data['downvote'])
-        db.session.add(vote)
-    elif not vote.upvote and not data['downvote']:
-        db.session.delete(vote)
-    elif vote.upvote and data['downvote']:
-        vote.upvote = False
-        db.session.add(vote)
-    
-    db.session.commit()
-    return Response(status=200)
