+++ reviveme/api/v1/thread_controller.py
@@ -80,51 +80,3 @@ def thread_delete(id):
     thread.deleted = True
     db.session.commit()
     return Response(status=200)
-
-@bp.route("/threads/<int:id>/upvote", methods=["POST"])
-def thread_upvote(id):
-    thread = db.get_or_404(Thread, id)
-    if thread.deleted:
-        return Response(status=400, response=f"Thread with id {id} has been deleted")
-    
-    data = UpVoteSchema().load(request.get_json())
-
-    vote = db.session.execute(
-        select(ThreadVote).where(ThreadVote.thread_id == id, ThreadVote.user_id == data['user_id'])
-    ).scalars().first()
-
-    if vote is None:
-        vote = ThreadVote(user_id=data['user_id'], thread_id=id, upvote=data['upvote'])
-        db.session.add(vote)
-    elif vote.upvote and not data['upvote']: # if user upvoted and now wants to remove upvote
-        db.session.delete(vote)
-    elif not vote.upvote and data['upvote']: # if user downvoted and now wants to upvote
-        vote.upvote = True
-        db.session.add(vote)
-    
-    db.session.commit()
-    return Response(status=200)
-
-@bp.route("/threads/<int:id>/downvote", methods=["POST"])
-def thread_downvote(id):
-    thread = db.get_or_404(Thread, id)
-    if thread.deleted:
-        return Response(status=400, response=f"Thread with id {id} has been deleted")
-    
-    data = DownVoteSchema().load(request.get_json())
-
-    vote = db.session.execute(
-        select(ThreadVote).where(ThreadVote.thread_id == id, ThreadVote.user_id == data['user_id'])
-    ).scalars().first()
-
-    if vote is None:
-        vote = ThreadVote(user_id=data['user_id'], thread_id=id, upvote=not data['downvote'])
-        db.session.add(vote)
-    elif not vote.upvote and not data['downvote']:
-        db.session.delete(vote)
-    elif vote.upvote and data['downvote']:
-        vote.upvote = False
-        db.session.add(vote)
-    
-    db.session.commit()
-    return Response(status=200)
